# api/openapi.yaml
openapi: 3.0.3
info:
  title: Warta API
  version: 1.0.0
  description: |
    REST API untuk platform artikel Warta: akun dan role, kategori, tag,
    artikel, dan komentar.

    **Autentikasi.** Login atau register menghasilkan `access_token` (JWT,
    berlaku singkat) dan `refresh_token` (berlaku lebih lama, sekali pakai).
    Kirim access token di header `Authorization: Bearer <token>`. Saat access
    token kedaluwarsa, tukar refresh token di `/auth/refresh` untuk mendapat
    pasangan baru. Refresh token lama langsung tidak berlaku; bila dipakai lagi,
    semua sesi akun itu dicabut.

    Endpoint publik tetap bisa dipanggil tanpa token. Namun token yang dikirim
    tapi tidak valid atau kedaluwarsa selalu ditolak dengan 401.

    **Role.** Akun baru selalu `reader` (bisa berkomentar). Admin menaikkannya
    menjadi `author` (bisa menulis artikel) atau `admin`.

    **Bentuk response.** Sukses: `{"data": ...}`, ditambah `meta` untuk daftar
    berhalaman. Gagal: `{"error": {"code", "message", "fields"}}`.
servers:
  - url: /
    description: Server yang sedang menyajikan dokumen ini

# Bawaannya token opsional; operasi yang wajib login menimpanya.
security:
  - {}
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Membuat akun baru (role reader)
      description: Dibatasi per IP, lihat `AUTH_RATE_LIMIT`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Akun dibuat dan langsung login
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Login dengan email dan password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Berhasil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Menukar refresh token dengan pasangan token baru
      description: |
        Refresh token lama langsung dicabut. Role di access token baru diambil
        dari database, jadi perubahan role oleh admin berlaku setelah refresh.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: Pasangan token baru
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Mencabut refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204": { description: "Berhasil, juga untuk token yang tidak dikenal" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/me:
    get:
      operationId: getMe
      tags: [Akun]
      summary: Data akun sendiri
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateMe
      tags: [Akun]
      summary: Mengubah nama
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/me/password:
    put:
      operationId: changePassword
      tags: [Akun]
      summary: Mengganti password
      description: Semua refresh token akun ini ikut dicabut.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "204": { description: Berhasil }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/me/articles:
    get:
      operationId: listMyArticles
      tags: [Akun]
      summary: Artikel milik sendiri, semua status
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/CategoryFilter"
        - $ref: "#/components/parameters/TagFilter"
        - $ref: "#/components/parameters/Sort"
        - name: status
          in: query
          schema: { type: string, enum: [draft, published, archived, all] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleListEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/me/bookmarks:
    get:
      tags: [Akun]
      operationId: listMyBookmarks
      summary: Artikel yang disimpan
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleListEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/stats:
    get:
      tags: [Akun]
      operationId: getStats
      summary: Statistik dashboard (author dan admin)
      description: Admin melihat semua artikel, author hanya artikelnya sendiri.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 7, maximum: 90, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Stats" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/uploads:
    post:
      tags: [Interaksi]
      operationId: uploadImage
      summary: Mengunggah gambar sampul (author dan admin)
      description: |
        JPEG, PNG, WebP, atau GIF, paling besar `MAX_UPLOAD_MB` (bawaan 2 MB).
        Format dikenali dari isi berkas, bukan dari nama atau header. Pakai
        `url` hasilnya sebagai `cover_image` artikel.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image: { type: string, format: binary }
      responses:
        "201":
          description: Tersimpan
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      url: { type: string, example: /uploads/3f2a9c0d4e5b6a7c8d9e0f1a2b3c4d5e.png }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/users:
    get:
      operationId: listUsers
      tags: [User]
      summary: Daftar user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: q
          in: query
          description: Cari di nama atau email
          schema: { type: string }
        - name: role
          in: query
          schema: { $ref: "#/components/schemas/Role" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
                  meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/users/{id}:
    get:
      operationId: getUser
      tags: [User]
      summary: Detail user
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [User]
      summary: Mengubah role user
      description: Admin tidak bisa mengubah role akunnya sendiri.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRoleRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/categories:
    get:
      operationId: listCategories
      tags: [Kategori]
      summary: Semua kategori
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Category" }
    post:
      operationId: createCategory
      tags: [Kategori]
      summary: Membuat kategori (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryRequest" }
      responses:
        "201":
          description: Dibuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/categories/{id}:
    get:
      operationId: getCategory
      tags: [Kategori]
      summary: Detail kategori berdasarkan id atau slug
      parameters: [{ $ref: "#/components/parameters/IdOrSlug" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateCategory
      tags: [Kategori]
      summary: Mengubah kategori (admin)
      description: Slug dibuat ulang bila nama berubah.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteCategory
      tags: [Kategori]
      summary: Menghapus kategori (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Dihapus }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Kategori masih dipakai artikel
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/tags:
    get:
      operationId: listTags
      tags: [Tag]
      summary: Daftar tag, diurutkan dari yang paling banyak dipakai
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: q
          in: query
          description: Cari di nama tag
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Tag" }
                  meta: { $ref: "#/components/schemas/PaginationMeta" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createTag
      tags: [Tag]
      summary: Membuat tag (admin)
      description: Author tidak perlu endpoint ini; tag baru dibuat otomatis saat menulis artikel.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagRequest" }
      responses:
        "201":
          description: Dibuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/tags/{id}:
    put:
      operationId: updateTag
      tags: [Tag]
      summary: Mengganti nama tag (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteTag
      tags: [Tag]
      summary: Menghapus tag dan melepasnya dari semua artikel (admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Dihapus }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/articles:
    get:
      operationId: listArticles
      tags: [Artikel]
      summary: Daftar artikel
      description: |
        Tanpa `status`, hanya artikel yang sudah terbit yang tampil. Status lain
        (`draft`, `archived`, `all`) hanya untuk admin. Daftar berisi cuplikan
        (`excerpt`), bukan isi lengkap.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/CategoryFilter"
        - $ref: "#/components/parameters/TagFilter"
        - name: author
          in: query
          description: Id penulis
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: string, enum: [draft, published, archived, all] }
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleListEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createArticle
      tags: [Artikel]
      summary: Menulis artikel (author atau admin)
      description: |
        Slug dibuat dari judul dan dibuat unik. Tag yang belum ada dibuat
        otomatis; nama tag dirapikan menjadi huruf kecil.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleRequest" }
      responses:
        "201":
          description: Dibuat
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api/v1/articles/{id}:
    get:
      operationId: getArticle
      tags: [Artikel]
      summary: Detail artikel berdasarkan id atau slug
      description: |
        Artikel yang belum terbit hanya terlihat oleh penulisnya dan admin.
        Bagi yang lain responsnya 404.
      security: [{}, { bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdOrSlug" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: replaceArticle
      tags: [Artikel]
      summary: Mengganti seluruh isi artikel (penulis atau admin)
      description: |
        Semua field wajib dikirim; `tags` yang tidak dikirim berarti artikel
        tidak punya tag. Slug mengikuti judul hanya selama artikel belum
        pernah terbit.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    patch:
      operationId: patchArticle
      tags: [Artikel]
      summary: Mengubah sebagian artikel (penulis atau admin)
      description: Hanya field yang dikirim yang diubah, misalnya `{"status":"published"}`.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticlePatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteArticle
      tags: [Artikel]
      summary: Menghapus artikel beserta komentarnya (penulis atau admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Dihapus }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/articles/{id}/like:
    put:
      tags: [Interaksi]
      operationId: likeArticle
      summary: Menyukai artikel terbit (idempoten)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Keadaan terbaru
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Engagement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Interaksi]
      operationId: unlikeArticle
      summary: Membatalkan suka
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Keadaan terbaru
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Engagement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/articles/{id}/bookmark:
    put:
      tags: [Interaksi]
      operationId: bookmarkArticle
      summary: Menyimpan artikel terbit (idempoten)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Keadaan terbaru
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Engagement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Interaksi]
      operationId: unbookmarkArticle
      summary: Menghapus dari simpanan
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Keadaan terbaru
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Engagement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/articles/{id}/view:
    post:
      tags: [Interaksi]
      operationId: recordArticleView
      summary: Mencatat artikel dibaca
      description: |
        Dipanggil halaman baca. Pembaca yang sama (akun, atau IP bila belum
        login) dihitung sekali per 30 menit, dan penulis yang membuka
        artikelnya sendiri tidak dihitung.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Tercatat atau diabaikan sebagai hitungan ganda }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/articles/{id}/comments:
    get:
      operationId: listComments
      tags: [Komentar]
      summary: Komentar sebuah artikel, dari yang terlama
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Comment" }
                  meta: { $ref: "#/components/schemas/PaginationMeta" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createComment
      tags: [Komentar]
      summary: Berkomentar (semua role, hanya di artikel yang sudah terbit)
      description: |
        Aturan anti-spam: paling banyak 2 tautan, tidak seluruhnya huruf
        kapital, komentar yang sama persis dalam 10 menit ditolak (409), dan
        tiap akun dibatasi `COMMENT_RATE_LIMIT` komentar per menit (429).
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentRequest" }
      responses:
        "201":
          description: Dibuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommentEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/v1/moderation/comments:
    get:
      operationId: listReportedComments
      tags: [Komentar]
      summary: Antrean moderasi (admin)
      description: Komentar yang dilaporkan atau disembunyikan, dari yang paling banyak dilaporkan.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ReportedComment" }
                  meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email, maxLength: 191 }
        password: { type: string, minLength: 8, description: Maksimal 72 byte }
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
...
    TokenEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/TokenPair" }
...